requirements.txt
requests
python-dateutil
pytz
python-dotenv
wordcloud
nltk
pandas
matplotlib.pyplot